+++ cli/src/main.rs
@@ -283,6 +283,15 @@ fn build_app<'a, 'b>() -> App<'a, 'b> {
                     ),
                 )
                 .subcommand(
+                    SubCommand::with_name("details")
+                        .about("Fetch details for a given ledger")
+                        .arg(
+                            Arg::with_name("sequence")
+                                .required(true)
+                                .help("The sequence of the ledger for which to fetch details")
+                        )
+                )
+                .subcommand(
                     listable!(
                         SubCommand::with_name("effects")
                             .about("Fetch effects for a given ledger")
@@ -430,6 +439,7 @@ fn main() {
         },
         ("ledgers", Some(sub_m)) => match sub_m.subcommand() {
             ("all", Some(sub_m)) => ledgers::all(&client, sub_m),
+            ("details", Some(sub_m)) => ledgers::details(&client, sub_m),
             ("effects", Some(sub_m)) => ledgers::effects(&client, sub_m),
             ("payments", Some(sub_m)) => ledgers::payments(&client, sub_m),
             ("operations", Some(sub_m)) => ledgers::operations(&client, sub_m),